timeout bugfix: allows the environment to finish its output message before raising the timeout event
Summary:
When a timeout is raised at exactly the specified time, it can break the bit encoding of the message
since it can kick in at any point in the middle of the serialization of a character. To avoid this, we allow
the environment to finish whatever message is outputting.

Differential Revision: D3505145

fbshipit-source-id: 504ef4a22858301a451bbe8497f983cad7c79b87