New helpers for handling the incoming message + refactoring of some related code + bugfix drawing world before initialization
Summary:
This mainly adds a couple of helper methods (get_match and is_message) to the MessageReceived event.
As a side effect, I decided to refactor a bit the code related to the events definition that was located in a not very appropriate place.
Morever, testing this I discovered a bug where the environment was crashing for trying to draw an unitialized world, so I fixed this.

Differential Revision: D3524490

fbshipit-source-id: a4df0fa8b614fd987c249d8e26c93576620cd995