task-specific tests
Summary:
Testing the tasks interactively with the human learner becomes a pain after a while, especially
if you have to go back to them when you make changes to the code. This is going in the direction of
having "pushing a button" tests for the tasks. The downside is that we will have to maintain the tests
when the tasks change, but I think they are at a mature enough level that we can start to afford that.

Differential Revision: D3594090

fbshipit-source-id: e10718815b0dff589752fc460b53f54ae026dee0