major refactoring moving the environment away from the tasks constructors
Summary:
I think it is a great idea to do a major refactoring of the code
two days before I leave. Move fast and break things, right? :))
The major refactoring consists in removing the environment from the constructor of the tasks.
Instead, the tasks receive the environment in the moment they are initialized at run time.
Also, now you can specify the kind of learner that you want to run through command line.

Differential Revision: D3662300

fbshipit-source-id: 7539ffde88cbed6059ff9028ca215450ce26eb26