implement non-interactive viewing mode
Summary:
This implements a non-interactive mode that is helpful for surveying the performance of the learners without interactivity. This is achieved by some refactoring + inheritance which minimizes duplicated code.
The user can choose the view, although if the learner is the HumanLearner, the view is the interactive ConsoleView by default.

Differential Revision: D3761065

fbshipit-source-id: 5580269601df0648c1fd673245e95b03596354b0