Support Learners implemented in any language
Summary:
This adds a `RemoteLearner` super class which allows the learner to be expressed as any arbitrary binary. The environment and learner communicate over a socket (fast).

If the learner is specified to be a `RemoteLearner`, the environment acts as a server. When run_tournament is executed, the environment awaits a handshake from a learner, and then they begin talking. For convenience, the user can specify the command to launch the learner so that it is launched with the environment in the same process.

The contract for a remote learner is:
1. send 'hello' to the environment
2. loop: accept reward (1 byte), accept env msg (1 byte), send reply (1 byte).

I've used a TARGETS file to compile the dependencies within fbcode, this is solved more easily when not working in fbcode.

Differential Revision: D3793543

fbshipit-source-id: 8069201192dcfc8d26cfd02f65e0c4422c57795d