Adding ManualHumanLearner (no automatic decisions as of what to output at each timestep), parametrizing serializer and fixing bug with is_message
Summary:
This adds some basic functionality to be able to debug the nano tasks. This adds the ability to set the serializer as follows:

-s core.serializer.IdentitySerializer

If you do the above, and you write your tasks only in terms of 1s and 0s, they will be translated directly into "bits" by the serializer
(just by passing them untouched as they come)

Furthermore, I have added the ManualHumanLearner that doesn't make any assumptions as to when to speak and basically asks for data at
every timestep.

Reviewed By: mbaroni

Differential Revision: D4543497

fbshipit-source-id: 026b5a5d755dad1cd4cee87bcb3eefa406f45139