Double bugfix. Bug introduced on the silence character generalization commit and subtle bug due to events being processed after the task has ended
Summary:
On the previous commit I introduced a bug by not allowing the set_reward to clear the output buffer (blocking impatient teachers kind of behavior). Furthermore, I noticed another subtle bug
produced by some events (in particular, on_output_message) being processed after the task has finished. So, for example, a learner may have failed the stay silent task. So the teacher sends a -1 reward
and starts producing a silence to separate to the next task (I know, this should go away in the near future). The new silence is what the teacher was expecting to reward the learner for completing the task
and now the learner gets a positive reward after all, even if it had failed before. The solution involves desconecting the event processing after the task has ended.

Reviewed By: mbaroni

Differential Revision: D4713062

fbshipit-source-id: 6fdcc6eed4f3fa6d5a9078b59bcad55264d5b701