Remove broken TARGETS files in fbcode/experimental
Summary:
We want to make it so that all TARGETS files in fbcode must be valid. This will help tools like pyre not fail on diffs. Previously this was impossible due to experimental... but with experimental going away, now it is realistic.

So I'm removing people's broken crap. Revert and fix it when you move out of experimental.

Differential Revision: D23587259

fbshipit-source-id: 19f6ca1dfba1587ef7c1b800c433eb56f35d0b4d